# delivery/http/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Personal Expense Tracker API
  description: |
    Personal Expense Tracker backend API.

    **Authentication flow**
    - Get `access_token` and `refresh_token` from `POST /auth/login`
    - Send `Authorization: Bearer <access_token>` to protected endpoints
    - Rotate tokens with `POST /auth/refresh`
    - Revoke the current refresh token with `POST /auth/logout`

    **Response format**
    - All endpoints return the same top-level envelope: `success`, `message`, `data`, `errors`, and `meta`
    - Validation and business errors always use `errors` as an array of strings
    - Paginated list endpoints return `data.items` and `meta.pagination`

    ## Features
    - User authentication and profile management
    - Expense tracking with categories
    - Debt tracking (lent/borrowed)
    - Spending reports and insights
    - Interactive API documentation
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /expenses:
    get:
      tags:
        - Expenses
      summary: List expenses
      description: List expenses for the current user. Optional filters by date range and category.
      security:
        - BearerAuth: []
      parameters:
        - name: from_date
          in: query
          description: Start date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          description: End date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: category_id
          in: query
          description: Filter by category UUID
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: List of expenses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseListResponse'
        '401':
          description: Missing or invalid authorization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Expenses
      summary: Create expense
      description: Create a new expense for the current user.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseRequest'
      responses:
        '201':
          description: Created expense
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseSuccessResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid authorization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /categories:
    get:
      tags:
        - Categories
      summary: List categories
      description: List global and user-defined categories for the authenticated user.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryListResponse'
        '400':
          description: Invalid pagination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid authorization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Categories
      summary: Create category
      description: Create a global or user-defined category.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201':
          description: Created category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategorySuccessResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /debts:
    get:
      tags:
        - Debts
      summary: List all debts
      description: Get all debts for the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: List of debts retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtListResponse'
        '400':
          description: Invalid pagination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid authorization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Debts
      summary: Create a new debt
      description: Add a new debt (lent or borrowed)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDebtInput'
            examples:
              lent_money:
                summary: Money lent to friend
                value:
                  type: "lent"
                  peer_name: "Alex"
                  amount: 100.00
                  due_date: "2024-02-01"
                  reminder_enabled: true
                  note: "Movie tickets"
              borrowed_money:
                summary: Money borrowed from friend
                value:
                  type: "borrowed"
                  peer_name: "Sarah"
                  amount: 50.00
                  due_date: "2024-02-15"
                  reminder_enabled: false
      responses:
        '201':
          description: Debt created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtSuccessResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid authorization
  /debts/upcoming:
    get:
      tags:
        - Debts
      summary: List upcoming debts
      description: Get debts due in the next X days (default 7)
      security:
        - BearerAuth: []
      parameters:
        - name: days
          in: query
          description: Number of days to look ahead
          required: false
          schema:
            type: integer
            default: 7
            minimum: 1
            example: 7
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: List of upcoming debts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtListResponse'
        '400':
          description: Invalid days parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid authorization
  /debts/{id}:
    put:
      tags:
        - Debts
      summary: Update a debt
      description: Update an existing debt (partial updates supported)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDebtInput'
            example:
              amount: 150.00
              note: "Updated amount"
      responses:
        '200':
          description: Debt updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebtSuccessResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Missing or invalid authorization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Debt not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    Page:
      name: page
      in: query
      description: Page number, starting from 1
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSize:
      name: page_size
      in: query
      description: Number of items per page
      required: false
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT token in the format "Bearer &lt;token&gt;"
  schemas:
    APIResponseBase:
      type: object
      description: Standard response envelope returned by all HTTP endpoints
      required:
        - success
        - message
        - data
        - errors
        - meta
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          nullable: true
        errors:
          type: array
          items:
            type: string
          nullable: true
        meta:
          nullable: true
    Error:
      description: Standard error response
      allOf:
        - $ref: '#/components/schemas/APIResponseBase'
        - type: object
          properties:
            success:
              type: boolean
              example: false
            message:
              type: string
              example: "Validation failed"
            data:
              nullable: true
              example: null
            errors:
              type: array
              items:
                type: string
              example:
                - "email is required"
            meta:
              nullable: true
              example: null
    Pagination:
      type: object
      description: Pagination metadata for list endpoints
      properties:
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 10
        total_items:
          type: integer
          example: 125
        total_pages:
          type: integer
          example: 13
        has_next:
          type: boolean
          example: true
        has_previous:
          type: boolean
          example: false
    Meta:
      type: object
      description: Optional metadata container
      properties:
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Expense:
      description: Expense resource
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        category_id:
          type: string
          format: uuid
          nullable: true
        is_recurring:
          type: boolean
        recurrence_type:
          type: string
          enum: [daily, weekly, monthly]
        next_due_date:
          type: string
          format: date
          nullable: true
        reminder_enabled:
          type: boolean
        reminder_sent_at:
          type: string
          format: date-time
          nullable: true
        note:
          type: string
        expense_date:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
    CreateExpenseRequest:
      type: object
      description: Expense creation payload
      required:
        - amount
        - expense_date
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: number
          format: double
          minimum: 0
        category_id:
          type: string
          format: uuid
          nullable: true
        is_recurring:
          type: boolean
          default: false
        recurrence_type:
          type: string
          enum: [daily, weekly, monthly]
        next_due_date:
          type: string
          format: date
        reminder_enabled:
          type: boolean
        note:
          type: string
        expense_date:
          type: string
          format: date
...
    ExpenseListData:
      type: object
      description: Paginated expense items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Expense'
    ExpenseSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/APIResponseBase'
        - type: object
          properties:
            success:
              type: boolean
              example: true
            message:
              type: string
              example: "Expense retrieved successfully"
            data:
              $ref: '#/components/schemas/Expense'
            errors:
              nullable: true
              example: null
            meta:
              nullable: true
              example: null
    ExpenseListResponse:
      allOf:
        - $ref: '#/components/schemas/APIResponseBase'
        - type: object
          properties:
            success:
              type: boolean
              example: true
            message:
              type: string
              example: "Expenses retrieved successfully"
            data:
              $ref: '#/components/schemas/ExpenseListData'
            errors:
              nullable: true
              example: null
            meta:
              $ref: '#/components/schemas/Meta'

    # ========================================
    # CATEGORY SCHEMAS (Team 2)
    # ========================================
    Category:
      description: Category resource
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        user_id:
          type: string
          format: uuid
          nullable: true
    CreateCategoryRequest:
      type: object
      description: Category creation payload
      required:
        - name
      properties:
        name:
          type: string
        user_id:
          type: string
          format: uuid
          nullable: true
...
    CategoryListData:
      type: object
      description: Paginated category items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Category'
    CategorySuccessResponse:
      allOf:
        - $ref: '#/components/schemas/APIResponseBase'
        - type: object
          properties:
            success:
              type: boolean
              example: true
            message:
              type: string
              example: "Category retrieved successfully"
            data:
              $ref: '#/components/schemas/Category'
            errors:
              nullable: true
              example: null
            meta:
              nullable: true
              example: null
    CategoryListResponse:
      allOf:
        - $ref: '#/components/schemas/APIResponseBase'
        - type: object
          properties:
            success:
              type: boolean
              example: true
            message:
              type: string
              example: "Categories retrieved successfully"
            data:
              $ref: '#/components/schemas/CategoryListData'
            errors:
              nullable: true
              example: null
            meta:
              $ref: '#/components/schemas/Meta'

    # ========================================
    # DEBT SCHEMAS
    # ========================================
    Debt:
      type: object
      description: Debt resource
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        user_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        type:
          type: string
          enum: [lent, borrowed]
          example: "lent"
        peer_name:
          type: string
          example: "Alex"
        amount:
          type: number
          format: float
          example: 100.50
        due_date:
          type: string
          format: date
          example: "2024-02-01"
        status:
          type: string
          enum: [pending, paid, overdue]
          example: "pending"
        reminder_enabled:
          type: boolean
          example: true
        note:
          type: string
          nullable: true
          example: "Movie tickets"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    CreateDebtInput:
      type: object
      description: Debt creation payload
      required:
        - type
        - peer_name
        - amount
        - due_date
      properties:
        type:
          type: string
          enum: [lent, borrowed]
          example: "lent"
        peer_name:
          type: string
          example: "Alex"
        amount:
          type: number
          format: float
          minimum: 0.01
          example: 100.50
        due_date:
          type: string
          format: date
          example: "2024-02-01"
        reminder_enabled:
          type: boolean
          default: false
          example: true
        note:
          type: string
          example: "Movie tickets"
    UpdateDebtInput:
      type: object
      description: Debt update payload
      properties:
        type:
          type: string
          enum: [lent, borrowed]
          example: "borrowed"
        peer_name:
          type: string
          example: "Sarah"
        amount:
          type: number
          format: float
          minimum: 0.01
          example: 150.00
        due_date:
          type: string
          format: date
          example: "2024-03-01"
        reminder_enabled:
          type: boolean
          example: false
        note:
          type: string
          example: "Updated note"
      minProperties: 1
    DebtListData:
      type: object
      description: Paginated debt items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Debt'
    DebtSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/APIResponseBase'
        - type: object
          properties:
            success:
              type: boolean
              example: true
            message:
              type: string
              example: "Debt retrieved successfully"
            data:
              $ref: '#/components/schemas/Debt'
            errors:
              nullable: true
              example: null
            meta:
              nullable: true
              example: null
    DebtListResponse:
      allOf:
        - $ref: '#/components/schemas/APIResponseBase'
        - type: object
          properties:
            success:
              type: boolean
              example: true
            message:
              type: string
              example: "Debts retrieved successfully"
            data:
              $ref: '#/components/schemas/DebtListData'
            errors:
              nullable: true
              example: null
            meta:
              $ref: '#/components/schemas/Meta'

    # ========================================
    # REPORT SCHEMAS
    # ========================================
...